+++ tests/opentest.py
@@ -29,7 +29,6 @@ import getopt
 import ctypes as ct
 
 import libpcap as pcap
-
 from pcaptestutils import *  # noqa
 
 #ifndef lint